+++ uv_hal/src/uv_utilities.c
@@ -279,7 +279,7 @@ float uv_lerpf(float t, float a, float b) {
 /// @param t: "Lerping scale". Should be between 0 ... 1000
 /// @param a: The value at t=0
 /// @param b: The value at t=1000
-float uv_lerpi(int t, int a, int b) {
+int uv_lerpi(int t, int a, int b) {
 	return ((1000-t)*a + t*b) / 1000;
 }
